requirements.txt
sympy==1.6.1
regex
requests
flask-cors
antlr4-python3-runtime>=4.7,<4.8
Django
djangorestframework
gunicorn
django-allow-cidr==0.5.0